Chore: Rename import file name due to the previous migration and modifications of directories and file names.
